avconv: only retry decoding on actual decoding errors
Errors during decoding are currently considered non-fatal and do not
terminate transcoding, so even if parts of the data are corrupted, the
rest may be decodable.

However, that should apply only to the actual decoding calls, not to the
failures elsewhere (e.g. configuring filters).

(cherry picked from commit 27085d1b47c3741cc0fac284c916127c4066d049)
Signed-off-by: Luca Barbato <lu_zero@gentoo.org>

Bug-Id: 1089

Thanks to Mingi Cho and Taekyoung Kwon of the Information Security Lab,
Yonsei University for reporting the issue on the release branch.